rid of MongoClientException as it interferes with com one

+++ mongo/src/main/java/db/client/mongo/gateway/repo/DeleteManyGateway.java
@@ -1,5 +1,6 @@
 package db.client.mongo.gateway.repo;
 
+import com.mongodb.MongoClientException;
 import com.mongodb.MongoException;
 import com.mongodb.MongoWriteConcernException;
 import com.mongodb.MongoWriteException;
@@ -11,6 +12,7 @@ import db.client.mongo.adapter.statement.DeleteManyAdoptedStatement;
 import db.client.mongo.gateway.contract.DBAwared;
 import db.client.mongo.gateway.contract.DeleteGateway;
 import db.client.mongo.gateway.result.QueryExecutionResultBuilderFacade;
+import db.client.mongo.validator.InvalidStatementException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 
@@ -26,12 +28,16 @@ public class DeleteManyGateway implements DeleteGateway {
 
 	@Override
 	public QueryExecutionResult delete(AdoptedStatement statement) {
+		if (!(statement instanceof DeleteManyAdoptedStatement)) throw new InvalidStatementException("Statement " +
+				statement.getClass().getSimpleName() + " is of wrong type");
 		DeleteManyAdoptedStatement deleteStatement = (DeleteManyAdoptedStatement) statement;
-		String collectionName = statement.getCollectionName();
-		MongoCollection collection = client.getCollection(collectionName);
 		try {
+			String collectionName = statement.getCollectionName();
+			MongoCollection collection = client.getCollection(collectionName);
 			DeleteResult deleteResult = collection.deleteMany(deleteStatement.getFilter());
 			return QueryExecutionResultBuilderFacade.deleteSuccesful(deleteResult);
+		} catch(MongoClientException e) {
+			return QueryExecutionResultBuilderFacade.deleteFailed(e);
 		} catch(MongoWriteException e) {
 			return QueryExecutionResultBuilderFacade.deleteFailed(e);
 		} catch(MongoWriteConcernException e) {
